头像

+++ src/main/java/com/education/service/serviceImpl/article/ArticleServiceImpl.java
@@ -53,7 +53,7 @@ public class ArticleServiceImpl implements ArticleService {
     private ElasticSearchService elasticSearchService;
 
     @Override
-    public void addArticle(ArticleDto articleDto) throws Exception{
+    public void addArticle(ArticleDto articleDto) throws Exception {
         //复制属性
         ArticleEntity articleEntity = new ArticleEntity();
         BeanUtils.copyProperties(articleDto, articleEntity);
@@ -64,7 +64,7 @@ public class ArticleServiceImpl implements ArticleService {
         articleMapper.insert(articleEntity);
 
         //判断是否启用,添加入es中
-        if (articleEntity.getStatus().equals(Constant.NUMBER_ONE)){
+        if (articleEntity.getStatus().equals(Constant.NUMBER_ONE)) {
             elasticSearchService.save(articleEntity);
         }
     }
@@ -118,7 +118,7 @@ public class ArticleServiceImpl implements ArticleService {
     }
 
     @Override
-    public void deleteArticleById(ArticleDto articleDto) throws Exception{
+    public void deleteArticleById(ArticleDto articleDto) throws Exception {
         //校验参数
         if (StringUtils.isBlank(articleDto.getId())) {
             throw new EducationException(ResultCode.FAILER_CODE.getCode(), "主键不能为空");
@@ -133,9 +133,14 @@ public class ArticleServiceImpl implements ArticleService {
         EntityUtil.addModifyInfo(articleEntity);
 
         articleMapper.updateById(articleEntity);
+        //判断是否启用,添加入es中
+        if (articleEntity.getStatus().equals(Constant.NUMBER_ONE)) {
+            elasticSearchService.save(articleEntity);
+        } else {
+            //删除es中数据
+            elasticSearchService.delete(articleEntity);
+        }
 
-        //删除es中数据
-        elasticSearchService.delete(articleEntity);
     }
 
     @Override
@@ -158,7 +163,7 @@ public class ArticleServiceImpl implements ArticleService {
     }
 
     @Override
-    public void updateArticle(ArticleDto articleDto) throws Exception{
+    public void updateArticle(ArticleDto articleDto) throws Exception {
         //校验数据
         ArticleEntity articleEntity = articleMapper.selectById(articleDto.getId());
         if (articleEntity == null) {
